`download_datasets.py` threw `FileNotFoundError`
the first call to download using `gdown` required the folder `./data` to exist otherwise a `FileNotFoundError` was thrown. I've just added an `os.makedirs()` that will create this folder. The `exist_ok=True` means that this should work even if the command is repeated (no error and won't overwrite existing folder).